W174s: the u8*/u8** template moves the EIGHT-TWIN mdlFile family in lockstep -- six more functions, one shared defect confirmed.
  0033BBE0, 0033C960, 0033D6D0, 0033E540, 0033F2E0   nd57 -> nd49 each
  00340100                                            nd73 -> nd65
  0033F6A0  232/224 nd154 -> 224/224 **MATCH**, window cleared
  003403B0  nd302 -> nd296 (dereferenced vertex-buffer pointer, sibling-derived loop/body ordering)
  00355DD0, 00356B00  MATCH; 00356C70 window cleared

The six-function lockstep move is decisive evidence about this family. These twins had already moved nd73 -> 57 together in an earlier wave; they have now moved again by exactly 8 points each on the u8*/u8** loaded-base recovery. **They share ONE remaining defect and it will close all of them at once.** That makes them collectively the single highest-value target left in the file -- six matches from one insight.

The u8*/u8** template has now been applied successfully in FOUR separate slices of this file by four different agents: the five-function VU family (nd8 -> MATCH), 00349870 and 0034A650 (nd8 -> MATCH), and now these six. It is the most reusable finding of the wave.

Also recorded: 00356C70's clear came from reconstructing mixed pointer/integer/float state roles, removing conversion glue, and correcting HIDDEN FLOAT call ABIs -- the hidden-argument class again, this time on the float side. And FUN_00318FC0's return/calls were canonicalized with a typed hidden-float ABI added for FUN_00282D40.

447 MATCH / 220 NONMATCHING, 667 unique markers, over-window down to 30 with no additions, pragma brackets 16/16, VU asm kernels kept contiguous and unchanged.

+++ src/h_cdvd.c
@@ -86,7 +86,7 @@ extern u32 D_007D4780;
 extern u32 D_007D4784;
 #pragma alias D_007D4780_abs D_007D4780
 extern u32 D_007D4780_abs[];
-extern u32 D_007D1F80[];
+extern u8 D_007D1F80[];
 extern u32 D_007CF780[];
 extern u32 D_007CDF38;
